pyproject.toml
[project]
name = "orchestra"
version = "0.4.0"
description = "AI Agent Orchestrator — beads task DAG, tmux sessions, Claude Code workers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.15.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "anthropic>=0.52.0",
]

[project.scripts]
orchestra = "orchestra.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
